drop dash in decomposeds3 driver name
