+++ pkg/storage/fs/decomposeds3/option.go
@@ -16,7 +16,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-package decomposed_s3
+package decomposeds3
 
 import (
 	"github.com/mitchellh/mapstructure"
