+++ pkg/storage/fs/loader/loader.go
@@ -23,7 +23,7 @@ import (
 	_ "github.com/opencloud-eu/reva/v2/pkg/ocm/storage/received"
 	_ "github.com/opencloud-eu/reva/v2/pkg/storage/fs/cephfs"
 	_ "github.com/opencloud-eu/reva/v2/pkg/storage/fs/decomposed"
-	_ "github.com/opencloud-eu/reva/v2/pkg/storage/fs/decomposed_s3"
+	_ "github.com/opencloud-eu/reva/v2/pkg/storage/fs/decomposeds3"
 	_ "github.com/opencloud-eu/reva/v2/pkg/storage/fs/eos"
 	_ "github.com/opencloud-eu/reva/v2/pkg/storage/fs/eosgrpc"
 	_ "github.com/opencloud-eu/reva/v2/pkg/storage/fs/eosgrpchome"
